Fix issue with no folder in KML file
If you export a single layer/folder from your google my map it is not put into a folder.  Current code does not check for no folders and only processes tracks and waypoint contained in folders.  So there needs to be special case added to handle the no folder case.

If you have a map with a single layer and export the entire map that single layer will be contained in a folder - so no issues with this case.

Also added and display counts for waypoint, tracks and folders.

Moved waypoint and track processing code into functions so it can be used again in no folder case.